// apps/web/lib/contracts.generated.ts
export const contracts = {
  "AttestiaProfileRegistry": {
    "address": "0x610b0399d82d8cE68E558a4f7a8CE87d84BdA32E",
    "abi": [
      {
        "type": "constructor",
        "inputs": [
          {
            "name": "initialOwner",
            "type": "address",
            "internalType": "address"
          },
          {
            "name": "identityRegistry_",
            "type": "address",
            "internalType": "address"
          }
        ],
        "stateMutability": "nonpayable"
      },
// ...
  },
// ...
} as const;